restructuracion de carpetas + insert de usuarios via Singup + tuto de pasos READMEOJADERUTA
