+++ views/login.pug
@@ -15,7 +15,7 @@ body
     a.icon(href='javascript:void(0);' onclick='myFunction()')
       i.fa.fa-bars
     .container
- form.small-form(action="/login", method="post")
+    form.small-form(action="/login", method="post")
             div.form
                 div.title Bienvenido 
                 .subtitle Logueate
